requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-tagging==0.4.6
idna==2.6
pytz==2017.3
requests==2.18.4
Unipath==1.1
urllib3==1.22
